Ran gen_code.php
The only changes here are inside comments. One is to whitespace, and the other is to spelling. There is no meaningful change from the previous version of the PR. This PR is compatible with TA-Lib 0.6.4.

I've verified this by going through the changes in the TA-Lib (in the previous commit) as well as compiling this and running the same tests as I previously ran. Those tests being validating that a pure PHP version of the new TA-Lib functions generate the same results when included in PECL Trader.